fixed movement

+++ classes/network.py
@@ -21,7 +21,7 @@ class Network:
     
     def read_pos(self, str):
         str = str.split(",")
-        return int(str[0]), int(str[1])
+        return float(str[0]), float(str[1])
 
     def make_pos(self,tup):
         return str(tup[0]) + "," + str(tup[1])
